+++ CHANGELOG.md
@@ -1,1 +1,6 @@
-TODO
+## 0.1.0 (23-08-2022):
+
+Initial release. Added few metrics:
+- Tls connections (`TlsConnectionsCollector`)
+- Tls handshakes (`TlsHandshakeMonitor`)
+- Kestrel connections (`KestrelMetricsCollector`)